Add abstract CardinalityConstraint
